LPS-163444 Move flex options to new component so we can reuse them for Collection Display
