+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/ContainerDisplayOptions.js
@@ -25,10 +25,7 @@ import {
 import selectSegmentsExperienceId from '../../../../../../app/selectors/selectSegmentsExperienceId';
 import updateItemConfig from '../../../../../../app/thunks/updateItemConfig';
 import {getLayoutDataItemPropTypes} from '../../../../../../prop-types/index';
-
-const ALIGN_ITEMS_STRETCH = 'align-items-stretch';
-const FLEX_WRAP_NOWRAP = 'flex-nowrap';
-const JUSTIFY_CONTENT_START = 'justify-content-start';
+import {FlexOptions} from './FlexOptions';
 
 const DISPLAY_OPTIONS = [
 	{
@@ -42,67 +39,6 @@ const DISPLAY_OPTIONS = [
 	{
 		label: Liferay.Language.get('flex-column'),
 		value: CONTENT_DISPLAY_OPTIONS.flexColumn,
-	},
-];
-
-const ALIGN_OPTIONS = [
-	{
-		label: Liferay.Language.get('start'),
-		value: 'align-items-start',
-	},
-	{
-		label: Liferay.Language.get('center'),
-		value: 'align-items-center',
-	},
-	{
-		label: Liferay.Language.get('end'),
-		value: 'align-items-end',
-	},
-	{
-		label: Liferay.Language.get('stretch'),
-		value: ALIGN_ITEMS_STRETCH,
-	},
-	{
-		label: Liferay.Language.get('baseline'),
-		value: 'align-items-baseline',
-	},
-];
-
-const FLEX_WRAP_OPTIONS = [
-	{
-		label: Liferay.Language.get('nowrap'),
-		value: FLEX_WRAP_NOWRAP,
-	},
-	{
-		label: Liferay.Language.get('wrap'),
-		value: 'flex-wrap',
-	},
-	{
-		label: Liferay.Language.get('wrap-reverse'),
-		value: 'flex-wrap-reverse',
-	},
-];
-
-const JUSTIFY_OPTIONS = [
-	{
-		label: Liferay.Language.get('start'),
-		value: JUSTIFY_CONTENT_START,
-	},
-	{
-		label: Liferay.Language.get('center'),
-		value: 'justify-content-center',
-	},
-	{
-		label: Liferay.Language.get('end'),
-		value: 'justify-content-end',
-	},
-	{
-		label: Liferay.Language.get('between'),
-		value: 'justify-content-between',
-	},
-	{
-		label: Liferay.Language.get('around'),
-		value: 'justify-content-around',
 	},
 ];
 
@@ -158,86 +94,20 @@ export default function ContainerDisplayOptions({item}) {
 			/>
 
 			{flexOptionsVisible && (
-				<>
-					<SelectField
-						field={{
-							label: Liferay.Language.get('flex-wrap'),
-							name: 'flexWrap',
-							typeOptions: {
-								validValues: FLEX_WRAP_OPTIONS,
-							},
-						}}
-						onValueSelect={(name, value) => {
-							dispatch(
-								updateItemConfig({
-									itemConfig: {
-										[name]:
-											value === FLEX_WRAP_NOWRAP
-												? ''
-												: value,
-									},
-									itemId: item.itemId,
-									segmentsExperienceId,
-								})
-							);
-						}}
-						value={item.config.flexWrap || FLEX_WRAP_NOWRAP}
-					/>
-
-					<div className="d-flex justify-content-between">
-						<SelectField
-							className="page-editor__sidebar__fieldset__field-small"
-							field={{
-								label: Liferay.Language.get('align-items'),
-								name: 'align',
-								typeOptions: {
-									validValues: ALIGN_OPTIONS,
+				<FlexOptions
+					itemConfig={item.config}
+					onConfigChange={(name, value) => {
+						dispatch(
+							updateItemConfig({
+								itemConfig: {
+									[name]: value,
 								},
-							}}
-							onValueSelect={(name, value) => {
-								dispatch(
-									updateItemConfig({
-										itemConfig: {
-											[name]:
-												value === ALIGN_ITEMS_STRETCH
-													? ''
-													: value,
-										},
-										itemId: item.itemId,
-										segmentsExperienceId,
-									})
-								);
-							}}
-							value={item.config.align || ALIGN_ITEMS_STRETCH}
-						/>
-
-						<SelectField
-							className="page-editor__sidebar__fieldset__field-small"
-							field={{
-								label: Liferay.Language.get('justify-content'),
-								name: 'justify',
-								typeOptions: {
-									validValues: JUSTIFY_OPTIONS,
-								},
-							}}
-							onValueSelect={(name, value) => {
-								dispatch(
-									updateItemConfig({
-										itemConfig: {
-											[name]:
-												value === JUSTIFY_CONTENT_START
-													? ''
-													: value,
-										},
-										itemId: item.itemId,
-										segmentsExperienceId,
-									})
-								);
-							}}
-							value={item.config.justify || JUSTIFY_CONTENT_START}
-						/>
-					</div>
-				</>
+								itemId: item.itemId,
+								segmentsExperienceId,
+							})
+						);
+					}}
+				/>
 			)}
 
 			<SelectField
